fix: export heading options (#295)
Moves `headingLevelOptions` to themeConfigOptions.tsx for export from
the package
Also updates a number of imports that should have been importing through
the package alias rather than specific files
