+++ packages/visual-editor/src/components/puck/Card.tsx
@@ -12,21 +12,16 @@ import {
   BasicSelector,
   BodyProps,
   getFontWeightOverrideOptions,
+  backgroundColors,
+  BackgroundStyle,
+  headingLevelOptions,
+  HeadingLevel,
 } from "../../index.js";
 import { Body } from "./atoms/body.js";
 import { CTA, CTAProps, linkTypeFields } from "./atoms/cta.js";
-import {
-  Heading,
-  HeadingProps,
-  HeadingLevel,
-  headingOptions,
-} from "./atoms/heading.js";
+import { Heading, HeadingProps } from "./atoms/heading.js";
 import { Section } from "./atoms/section.js";
 import { imageWrapperVariants, ImageWrapperProps } from "./Image.js";
-import {
-  backgroundColors,
-  BackgroundStyle,
-} from "../../utils/themeConfigOptions.js";
 
 const PLACEHOLDER_IMAGE_URL = "https://placehold.co/640x360";
 
@@ -88,7 +83,7 @@ const cardFields: Fields<CardProps> = {
         { value: "uppercase", label: "Uppercase" },
         { value: "capitalize", label: "Capitalize" },
       ]),
-      level: BasicSelector("Level", headingOptions),
+      level: BasicSelector("Level", headingLevelOptions),
     },
   },
   subheading: {
@@ -109,7 +104,7 @@ const cardFields: Fields<CardProps> = {
         { value: "uppercase", label: "Uppercase" },
         { value: "capitalize", label: "Capitalize" },
       ]),
-      level: BasicSelector("Level", headingOptions),
+      level: BasicSelector("Level", headingLevelOptions),
     },
   },
   body: {
