+++ packages/visual-editor/src/components/puck/Footer.tsx
@@ -7,6 +7,8 @@ import {
   themeManagerCn,
   useDocument,
   BasicSelector,
+  type BackgroundStyle,
+  backgroundColors,
 } from "../../index.ts";
 import {
   FaFacebook,
@@ -17,10 +19,6 @@ import {
   FaTwitter,
   FaYoutube,
 } from "react-icons/fa";
-import {
-  type BackgroundStyle,
-  backgroundColors,
-} from "../../utils/themeConfigOptions.ts";
 
 type socialLink = {
   name: string;
