+++ packages/visual-editor/src/utils/index.ts
@@ -18,6 +18,8 @@ export {
   backgroundColors,
   defaultThemeTailwindExtensions,
   type BackgroundStyle,
+  headingLevelOptions,
+  type HeadingLevel,
   letterSpacingOptions,
   textTransformOptions,
 } from "./themeConfigOptions.ts";
