+++ packages/visual-editor/src/utils/themeConfigOptions.ts
@@ -88,6 +88,19 @@ export const backgroundColors: Record<string, BackgroundOption> = {
   },
 };
 
+// Defines the valid levels for the heading element
+export type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6;
+
+// Provides a mapping of label to value for BasicSelector
+export const headingLevelOptions = [
+  { label: "H1", value: 1 },
+  { label: "H2", value: 2 },
+  { label: "H3", value: 3 },
+  { label: "H4", value: 4 },
+  { label: "H5", value: 5 },
+  { label: "H6", value: 6 },
+];
+
 // Tailwind Theme Extensions (https://v3.tailwindcss.com/docs/theme#extending-the-default-theme)
 // to use in the tailwind.config.ts in conjunction with themeResolver and the theme.config
 export const defaultThemeTailwindExtensions = {
